+++ src/os2datascanner/projects/admin/adminapp/views/views.py
@@ -11,9 +11,11 @@
 # OS2datascanner is developed by Magenta in collaboration with the OS2 public
 # sector open source network <https://os2.eu/>.
 #
+import csv
+
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.forms.models import modelform_factory
-from django.http import Http404
+from django.http import Http404, StreamingHttpResponse
 from django.shortcuts import get_object_or_404
 from django.utils.translation import gettext_lazy as _
 from django.views.generic import ListView, TemplateView, DetailView
@@ -21,6 +23,7 @@ from django.views.generic.edit import CreateView, UpdateView
 from django.views.generic.edit import ModelFormMixin, DeleteView
 from django.conf import settings
 
+from os2datascanner.utils.system_utilities import time_now
 from os2datascanner.projects.admin.utilities import UserWrapper
 from ..models.scannerjobs.dropboxscanner import DropboxScanner
 from ..models.scannerjobs.exchangescanner import ExchangeScanner
@@ -210,3 +213,43 @@ class DialogSuccess(TemplateView):
             model_type = self.reload_map[model_type]
         context['reload_url'] = '/' + model_type + '/'
         return context
+
+
+class CSVExportMixin:
+    """View mixin for exporting a queryset normally delivered to a template
+    as a CSV-file instead. Intended use: Define a new view, which inherits
+    from the view, which normally delivers context to a template, and this
+    mixin. It is important, that the new view inherits from this mixin first!"""
+    paginator_class = None  # We never want to paginate results
+    exported_fields = {}  # Structure: {"<label>": "<field_name>", ...}
+    exported_filename = "exported_file"
+
+    class CSVBuffer:
+        def write(self, value):
+            return value
+
+    def stream_queryset(self, queryset):
+        """Writes to a virtual buffer, so there is only ever one row of the CSV
+        in memory."""
+        pseudo_buffer = self.CSVBuffer()
+        writer = csv.writer(pseudo_buffer)
+
+        yield writer.writerow(self.exported_fields.keys())
+
+        for obj in queryset:
+            yield writer.writerow(obj.values())
+
+    def get(self, request, *args, **kwargs):
+        # Since we are streaming, we need to select the entire queryset, and
+        # stream it from memory. We are not able to make further queries after
+        # streaming has begun.
+        queryset = list(self.get_queryset().values(*self.exported_fields.values()))
+
+        response = StreamingHttpResponse(
+            self.stream_queryset(queryset),
+            content_type="text/csv",
+            headers={
+                "Content-Disposition":
+                f'attachment; filename="{time_now()}-{self.exported_filename}.csv"'})
+
+        return response